+++ etc/webapi.conf
@@ -121,7 +121,7 @@ Nickname = "cn"
 Phone = "mobile"
 Email = "mail"
 
-[SSO]
+[OIDC]
 Enable = false
 RedirectURL = "http://n9e.com/callback"
 SsoAddr = "http://sso.example.org"
@@ -130,7 +130,7 @@ ClientSecret = ""
 CoverAttributes = true
 DefaultRoles = ["Standard"]
 
-[SSO.Attributes]
+[OIDC.Attributes]
 Nickname = "nickname"
 Phone = "phone_number"
 Email = "email"
