+++ src/webapi/webapi.go
@@ -13,7 +13,7 @@ import (
 	"github.com/didi/nightingale/v5/src/pkg/httpx"
 	"github.com/didi/nightingale/v5/src/pkg/ldapx"
 	"github.com/didi/nightingale/v5/src/pkg/logx"
-	"github.com/didi/nightingale/v5/src/pkg/ssoc"
+	"github.com/didi/nightingale/v5/src/pkg/oidcc"
 	"github.com/didi/nightingale/v5/src/storage"
 	"github.com/didi/nightingale/v5/src/webapi/config"
 	"github.com/didi/nightingale/v5/src/webapi/prom"
@@ -92,7 +92,7 @@ func (a Webapi) initialize() (func(), error) {
 	ldapx.Init(config.C.LDAP)
 
 	// init ssoc
-	ssoc.Init(config.C.SSO)
+	oidcc.Init(config.C.OIDC)
 
 	// init logger
 	loggerClean, err := logx.Init(config.C.Log)
